Removed Tk() and Toplevel() from Application
To unify the Form object, Tk() has been removed from Application.
This was originally to allow the messagebox module to elide the
creation of a Tk window when the "app" instance was never run.
The tkMessageBox library requires message boxes to be called via
a Tk widget so if one does not exist, the underlying library
creates one for you. Form.check_handle() is called when a message
box function is called and promptly creates and destroys a Tk
window so that a message box function in tkMessageBox can be called
without displaying a form, if one has not been created before
invoking a message box function.