Removed unused import from messagebox module.
Cleaned up some code in the Form class.

And mucked around with the tkplus test class.  This will eventually
go away and be replaced with some demos in a separate repo.  This
file is only here for testing purposes and to add a smoke test for
basic functionality testing of each widget.